Перенос логики анимации аватара
